An workaround for the include path issue on Mac OS X
In Mac OS X, clang appends an unwanted string '(framework directory)'
at the end of some include directories. This makes the plugin not working.

This workaround patch removes the extra suffix '(framework directory)'
from the include path. It seems a little bit dirty, but it would be
very rare for the include directory to contain it at the end.