Preserve haslocaldir() flag of windows.
Incautious use of lcd in various functions forced C/C++ source windows
to switch to localdir mode even when they previously were following the
global working directory.

While crudely so this fix makes sure the correct command is used when
restoring the previous working directory so that the haslocaldir() flag
is also restored along with it.