Handle a 'arguments' key in a compilation database.
A tool such as `bear` (https://github.com/rizsotto/Bear) uses the
`arguments` key instead of the `command` key in case of the compilation
database that is build for a Makefile project. According to
https://clang.llvm.org/docs/JSONCompilationDatabase.html#format either
`arguments` or `command` is required so this PR adds a support also for
the `arguments` key.