Use the old CWD when resetting the CWD
Currently, if the source window uses a localdir, we return the working directory
to that local dir.

However, if the source window uses the global working directory, we changed the
global directory to the user's homedire (because getcwd(-1) returns an empty
string).

Instead, we return the global working directory to the previous result of
getcwd()