Use absolute path for dotclang files
Currently, there's a bug where if the dotclang file sits in the same location as
the file, findfile returns just the filename rather than the absolute path.

Because of this, once we return to the previous working directory - we can't
find the .clang file anymore and hence we ignore it.

We fix this by always taking the absolute path of whatever findfile returns.